// Nordax.Bank.Recruitment/ClientApp/src/models/newLoanApplicationRequest.ts
export interface NewLoanApplicationRequest {
    loanAmount: number;
    loanDurationYears: number;
    monthlyIncome: number;
    companyName: string;
    employmentType: string;
    employedSince: number;
    personalNumber: string;
    firstName: string;
    lastName: string;
    email: string;
    phoneNumber: number;
}